The base working version of the project
I had created this one a while back, so there weren't any steps between creation of the repository and publishing. I have several other's that will be like this, but I'll start to have normal workflow when I upload all of my favorites and have to start making new projects.